Refactor CellError and SheetError in log package
