+++ internal/log/error.go
@@ -1,4 +1,4 @@
-package sheet
+package log
 
 import "fmt"
 
@@ -8,4 +8,12 @@ func CellErrorOf(err error, sheetName, cellName string) error {
 	}
 
 	return fmt.Errorf("%w: in %q at %q", err, sheetName, cellName)
+}
+
+func SheetErrorOf(err error, sheetName string) error {
+	if err == nil {
+		return nil
+	}
+
+	return fmt.Errorf("%w: in %q", err, sheetName)
 }
