+++ internal/sheet/manager.go
@@ -589,7 +589,7 @@ func loadMembers(t *table.Table) *store.MemberStore {
 				Name:       strings.TrimSpace(cells[0].Value),
 				CardNumber: strings.TrimSpace(cells[1].Value),
 			})
-			fatalIfNotNil(CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 0)))
+			fatalIfNotNil(log.CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 0)))
 		},
 		IncludeHeader:   false,
 		UnknownRowCount: true,
@@ -614,7 +614,7 @@ func loadExpenses(t *table.Table, members *store.MemberStore) []*model.Expense {
 			}
 
 			theTime, timeErr := model.ParseTime(cells[0].Value)
-			fatalIfNotNil(CellErrorOf(timeErr, t.SheetName, t.GetCell(rowNumber, 0)))
+			fatalIfNotNil(log.CellErrorOf(timeErr, t.SheetName, t.GetCell(rowNumber, 0)))
 
 			title := cells[1].Value
 
@@ -622,7 +622,7 @@ func loadExpenses(t *table.Table, members *store.MemberStore) []*model.Expense {
 			requireMemberPresence(members, payer)
 
 			amount, amountErr := model.ParseAmount(cells[3].Value)
-			fatalIfNotNil(CellErrorOf(amountErr, t.SheetName, t.GetCell(rowNumber, 3)))
+			fatalIfNotNil(log.CellErrorOf(amountErr, t.SheetName, t.GetCell(rowNumber, 3)))
 
 			ex := &model.Expense{
 				Title:     title,
@@ -634,7 +634,7 @@ func loadExpenses(t *table.Table, members *store.MemberStore) []*model.Expense {
 			var shares []model.Share
 			for i := 4; i < t.ColumnCount; i += 2 {
 				weight, err := model.ParseShareWeight(cells[i].Value)
-				fatalIfNotNil(CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, i)))
+				fatalIfNotNil(log.CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, i)))
 
 				memberName := members.RequireMemberByIndex((i - 4) / 2).Name
 				shares = append(shares, model.Share{
@@ -659,7 +659,7 @@ func loadTransactions(t *table.Table, members *store.MemberStore) []*model.Trans
 	t.ReadRows(table.ReadRowsParams{
 		RowReader: func(rowNumber int, cells []*table.RCell) {
 			theTime, err := model.ParseTime(cells[0].Value)
-			fatalIfNotNil(CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 0)))
+			fatalIfNotNil(log.CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 0)))
 
 			receiver := cells[1].Value
 			requireMemberPresence(members, receiver)
@@ -668,7 +668,7 @@ func loadTransactions(t *table.Table, members *store.MemberStore) []*model.Trans
 			requireMemberPresence(members, payer)
 
 			amount, err := model.ParseAmount(cells[3].Value)
-			fatalIfNotNil(CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 3)))
+			fatalIfNotNil(log.CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, 3)))
 
 			transactions = append(transactions, &model.Transaction{
 				Time:         theTime,
@@ -700,7 +700,7 @@ func loadBaseState(t *table.Table, members *store.MemberStore) [][]model.Amount 
 
 			for i := 0; i < members.Count(); i++ {
 				amount, err := model.ParseAmount(cells[i+1].Value)
-				fatalIfNotNil(CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, i+1)))
+				fatalIfNotNil(log.CellErrorOf(err, t.SheetName, t.GetCell(rowNumber, i+1)))
 				baseState[rowNumber][i] = amount
 			}
 		},
